% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supabase_storage.R
\name{build_list_url}
\alias{build_list_url}
\title{Build Supabase Storage List API URL}
\usage{
build_list_url(conn, path = "")
}
\arguments{
\item{conn}{Supabase connection object}

\item{path}{Path within the bucket}
}
\value{
Full URL for list API
}
\description{
Build Supabase Storage List API URL
}
\keyword{internal}
