% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supabase_storage.R
\name{guess_content_type}
\alias{guess_content_type}
\title{Guess MIME content type from file extension}
\usage{
guess_content_type(file_path)
}
\arguments{
\item{file_path}{Path to file}
}
\value{
MIME type string
}
\description{
Guess MIME content type from file extension
}
\keyword{internal}
