% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studio.R
\name{sd_studio}
\alias{sd_studio}
\title{Launch surveydown Studio}
\usage{
sd_studio(gssencmode = "prefer")
}
\arguments{
\item{gssencmode}{Character string. The GSS encryption mode for the database
connection. Defaults to \code{"prefer"}. Set to \code{"disable"} if you're having
connection issues on a secure connection like a VPN.}
}
\value{
No return value, called for its side effects of launching a Shiny app.
}
\description{
This function launches a Shiny app with 4 tabs: Build, Preview, Responses, and Settings.
The Build tab includes a template selection interface for creating new surveys.
}
\examples{
if (interactive()) {
  # Launch studio
  sd_studio()
  
  # Launch studio with disabled GSS encryption (for VPN connections)
  sd_studio(gssencmode = "disable")
}
}
